+++ Game/HeadRollerStates.cpp
@@ -356,7 +356,7 @@ void HeadRollerDieState::Tick()
 		return;
 	}
 
-	float frameTime = animInstance->GetTotalFrame();
+	float frameTime = static_cast<float>(animInstance->GetTotalFrame());
 	frameTime /= 25;
 	m_fDissolveTimer += TIMER->GetDeltaTime();
 	float t = m_fDissolveTimer / frameTime;
