+++ Game/MageStates.cpp
@@ -581,7 +581,7 @@ void MageDieState::Tick()
 		End();
 	}
 
-	float frameTime = animInstance->GetTotalFrame();
+	float frameTime = static_cast<float>(animInstance->GetTotalFrame());
 	frameTime /= 20;
 	m_fDissolveTimer += TIMER->GetDeltaTime();
 	float t = m_fDissolveTimer / frameTime;
