+++ Game/WalkerStates.cpp
@@ -141,7 +141,7 @@ void WalkerDieState::Tick()
 		End();
 	}
 
-	float frameTime = animInstance->GetTotalFrame();
+	float frameTime = static_cast<float>(animInstance->GetTotalFrame());
 	frameTime /= 30;
 	m_fDissolveTimer += TIMER->GetDeltaTime();
 	float t = m_fDissolveTimer / frameTime;
